ZEGGS/requirements.txt
Flask==3.1.0
matplotlib==3.10.0
numpy==1.22.4
omegaconf==2.1.1
pandas==1.3.3
pyloudnorm==0.1.0
python-dotenv==1.0.1
python_dateutil==2.8.2
rich==13.9.4
scipy==1.7.1
sox==1.4.1
torch==2.1.2
